+++ IMS2/Controllers/DepartmentController.cs
@@ -8,6 +8,8 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using IMS2.Models;
+using IMS2.ViewModels;
+using System.Data.Entity.Infrastructure;
 
 namespace IMS2.Controllers
 {
@@ -16,10 +18,18 @@ namespace IMS2.Controllers
         private ImsDbContext db = new ImsDbContext();
 
         // GET: Department
-        public async Task<ActionResult> Index()
+        public async Task<ActionResult> Index(IMSMessageIdEnum? message)
         {
+            ViewBag.StatusMessage =
+               message == IMSMessageIdEnum.CreateSuccess ? "已创建新项。"
+               : message == IMSMessageIdEnum.EditdSuccess ? "已更新完成。"
+               : message == IMSMessageIdEnum.DeleteSuccess ? "已删除成功。"
+               : message == IMSMessageIdEnum.CreateError ? "创建项目出现错误。"
+               : message == IMSMessageIdEnum.EditError ? "有重名，无法更新相关信息。"
+               : message == IMSMessageIdEnum.DeleteError ? "不允许删除该项。"
+               : "";
             var departments = db.Departments.Include(d => d.DepartmentCategory);
-            return View(await departments.OrderBy(d=>d.Priority).ToListAsync());
+            return View(await departments.OrderBy(d => d.Priority).ToListAsync());
         }
 
         // GET: Department/Details/5
@@ -53,10 +63,19 @@ namespace IMS2.Controllers
         {
             if (ModelState.IsValid)
             {
-                department.DepartmentId = Guid.NewGuid();
-                db.Departments.Add(department);
-                await db.SaveChangesAsync();
-                return RedirectToAction("Index");
+                var query = await db.Departments.Where(d => d.DepartmentId == department.DepartmentId || d.DepartmentName == department.DepartmentName)
+                            .SingleOrDefaultAsync();
+                if (query == null)
+                {
+                    db.Departments.Add(department);
+                    await db.SaveChangesAsync();
+                    return RedirectToAction("Index", new { message = IMSMessageIdEnum.CreateSuccess });
+
+                }
+                else
+                {
+                    return RedirectToAction("Index", new { message = IMSMessageIdEnum.CreateError });
+                }
             }
 
             ViewBag.DepartmentCategoryId = new SelectList(db.DepartmentCategories, "DepartmentCategoryId", "DepartmentCategoryName", department.DepartmentCategoryId);
@@ -88,10 +107,44 @@ namespace IMS2.Controllers
         {
             if (ModelState.IsValid)
             {
-                db.Entry(department).State = EntityState.Modified;
-                await db.SaveChangesAsync();
-                return RedirectToAction("Index");
+                if (TryUpdateModel(department, "", new string[] { "DepartmentCategoryId", "DepartmentName", "Priority", "Remarks" }))
+                {
+                    var query = await db.Departments.Where(d => d.DepartmentName == department.DepartmentName).FirstOrDefaultAsync();
+                    if (query != null)
+                    {
+                        ModelState.AddModelError("", String.Format("已有科室名：{0}", department.DepartmentName));
+
+                    }
+                    else
+                    {
+                        db.Entry(department).State = EntityState.Modified;
+                        //client win
+                        bool saveFailed;
+                        do
+                        {
+                            saveFailed = false;
+                            try
+                            {
+                                await db.SaveChangesAsync();
+
+                            }
+                            catch (DbUpdateConcurrencyException ex)
+                            {
+                                saveFailed = true;
+
+                                // Update original values from the database 
+                                var entry = ex.Entries.Single();
+                                entry.OriginalValues.SetValues(entry.GetDatabaseValues());
+                            }
+
+                        } while (saveFailed);
+
+                        return RedirectToAction("Index", new { message = IMSMessageIdEnum.EditdSuccess });
+                    }
+                    
+                }
             }
+
             ViewBag.DepartmentCategoryId = new SelectList(db.DepartmentCategories, "DepartmentCategoryId", "DepartmentCategoryName", department.DepartmentCategoryId);
             return View(department);
         }
@@ -117,9 +170,18 @@ namespace IMS2.Controllers
         public async Task<ActionResult> DeleteConfirmed(Guid id)
         {
             Department department = await db.Departments.FindAsync(id);
-            db.Departments.Remove(department);
-            await db.SaveChangesAsync();
-            return RedirectToAction("Index");
+            if( department.DepartmentIndicatorValues.Count <= 0
+                && department.DepartmentIndicatorStandards.Count <= 0
+                && department.ProvidingIndicators.Count <= 0)
+            {
+                db.Departments.Remove(department);
+                await db.SaveChangesAsync();
+                return RedirectToAction("Index", new { message = IMSMessageIdEnum.DeleteSuccess });
+            }
+            else
+            {
+                return RedirectToAction("Index", new { message = IMSMessageIdEnum.DeleteError });
+            }
         }
 
         protected override void Dispose(bool disposing)
